+++ pydl/photoop/__init__.py
@@ -14,7 +14,7 @@ class PhotoopException(Exception):
 #
 # Import subpackages
 #
-#import photoobj
+import sdssio
 #
 # Set __all__
 #