+++ pydl/photoop/sdssio/__init__.py
@@ -1,0 +1,7 @@
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+# -*- coding: utf-8 -*-
+"""
+This subpackage corresponds to the sdssio directory of photoop.
+"""
+from filtername import filtername
+from filternum import filternum