+++ pydl/pydlutils/__init__.py
@@ -1,7 +1,7 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 # -*- coding: utf-8 -*-
 """
-This subpackage implements functions from the photoop package.
+This subpackage implements functions from the idlutils package.
 """
 #
 # Define this early on so that submodules can use it